revert sidebar layout
