--- frontend/src/lib/sidebar.ts
@@ -1,4 +1,0 @@
-import { writable } from 'svelte/store';
-
-/** controls whether the sidebar is open on small screens */
-export const sidebarOpen = writable(false);
